Cleanup, support CheckerInfoMessage

+++ EnoCore/Models/Json/JsonConfigurationService.cs
@@ -7,12 +7,12 @@ namespace EnoCore.Models.Json
     public class JsonConfigurationService
     {
 #pragma warning disable CS8618
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
-        public int FlagsPerRound { get; set; }
-        public int NoisesPerRound { get; set; }
-        public int HavocsPerRound { get; set; }
-        public int WeightFactor { get; set; }
+        public long FlagsPerRound { get; set; }
+        public long NoisesPerRound { get; set; }
+        public long HavocsPerRound { get; set; }
+        public long WeightFactor { get; set; }
         public bool Active { get; set; }
         public string[] Checkers { get; set; }
 #pragma warning restore CS8618
